editor refactor
